change decimal to double. finish refactoring filter attributes change dto files add validation to database

+++ Contracts/ICurrencyApiConnection.cs
@@ -2,6 +2,6 @@
 {
     public interface ICurrencyApiConnection
     {
-        public decimal GetExchangeRate(string currencyName);
+        public double GetExchangeRate(string currencyName);
     }
 }
